FootballCompetition-main/requirements.txt
requests==2.31.0
mysql-connector-python==8.2.0
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
pytest==7.4.3
docker==6.1.3
